generic

+++ everybodycodes/2025/src/util/file.rs
@@ -1,6 +1,6 @@
 use std::{env, fs};
 
-pub fn input(day: i32, part: i32) -> String {
+pub fn input(day: i8, part: i8) -> String {
     let demo = env::args().any(|a| a == "test");
     let input = if demo {
         format!("test-input/{}.txt", day)
